+++ OzzCodeGen/CodeEngines/CsSqliteRepository/SqliteRepositoryEntitySetting.cs
@@ -1,20 +1,46 @@
+using OzzCodeGen.CodeEngines.Storage;
+using OzzUtils;
 using System.Linq;
 
 namespace OzzCodeGen.CodeEngines.CsSqliteRepository;
 
 public class SqliteRepositoryEntitySetting : CsModelClass.BaseModelClassEntitySetting<SqliteRepositoryPropertySetting>
 {
+    public StorageEntitySetting StorageEntitySetting
+    {
+        get
+        {
+            CSharpSqliteRepositoryEngine codeEngine = CodeEngine as CSharpSqliteRepositoryEngine;
+            if (_storageEntitySetting == null && codeEngine?.StorageCodeEngine != null)
+            {
+                _storageEntitySetting = codeEngine.StorageCodeEngine.Entities.FirstOrDefault(e => e.Name == Name);
+            }
+
+            return _storageEntitySetting;
+        }
+    }
+    private StorageEntitySetting _storageEntitySetting;
+
     public string TableName
     {
         get
         {
-            if (string.IsNullOrEmpty(_tableName))
-                _tableName = Name;
+            if (string.IsNullOrEmpty(_tableName) && StorageEntitySetting != null)
+            {
+                _tableName = StorageEntitySetting.TableName;
+            }
+            else if (string.IsNullOrEmpty(_tableName))
+            {
+                _tableName = Name.Pluralize();
+            }
+
             return _tableName;
         }
         set
         {
-            if (_tableName == value) return;
+            if (_tableName == value)
+                return;
+
             _tableName = value;
             RaisePropertyChanged(nameof(TableName));
         }